+++ src/main/kotlin/net/liquidev/dawd3/render/Sprite.kt
@@ -1,8 +1,8 @@
 package net.liquidev.dawd3.render
 
 data class Sprite(
-    val u: Int,
-    val v: Int,
-    val width: Int,
-    val height: Int,
+    val u: Float,
+    val v: Float,
+    val width: Float,
+    val height: Float,
 )
