+++ src/main/kotlin/net/liquidev/dawd3/ui/widget/keyboard/Key.kt
@@ -10,33 +10,33 @@ import net.liquidev.dawd3.ui.widget.Widget
 import net.minecraft.client.util.math.MatrixStack
 
 class Key(
-    x: Int,
-    y: Int,
+    x: Float,
+    y: Float,
     val type: Type,
     val note: Float,
     val pitchControl: FloatControl,
 ) : Widget(x, y) {
 
     enum class Type(
-        val width: Int,
-        val height: Int,
+        val width: Float,
+        val height: Float,
         val idleSprite: Sprite,
         val pressedSprite: Sprite,
         val currentPitchSprite: Sprite,
     ) {
         White(
-            width = 10,
-            height = 32,
-            idleSprite = Sprite(u = 32, v = 16, width = 10, height = 32),
-            pressedSprite = Sprite(u = 42, v = 16, width = 10, height = 32),
-            currentPitchSprite = Sprite(u = 59, v = 16, width = 4, height = 4),
+            width = 10f,
+            height = 32f,
+            idleSprite = Sprite(u = 32f, v = 16f, width = 10f, height = 32f),
+            pressedSprite = Sprite(u = 42f, v = 16f, width = 10f, height = 32f),
+            currentPitchSprite = Sprite(u = 59f, v = 16f, width = 4f, height = 4f),
         ),
         Black(
-            width = 7,
-            height = 20,
-            idleSprite = Sprite(u = 52, v = 16, width = 7, height = 20),
-            pressedSprite = Sprite(u = 52, v = 36, width = 7, height = 20),
-            currentPitchSprite = Sprite(u = 59, v = 20, width = 3, height = 3),
+            width = 7f,
+            height = 20f,
+            idleSprite = Sprite(u = 52f, v = 16f, width = 7f, height = 20f),
+            pressedSprite = Sprite(u = 52f, v = 36f, width = 7f, height = 20f),
+            currentPitchSprite = Sprite(u = 59f, v = 20f, width = 3f, height = 3f),
         ),
     }
 
@@ -45,18 +45,23 @@ class Key(
 
     private var isPressed = false
 
-    override fun drawContent(matrices: MatrixStack, mouseX: Int, mouseY: Int, deltaTime: Float) {
+    override fun drawContent(
+        matrices: MatrixStack,
+        mouseX: Float,
+        mouseY: Float,
+        deltaTime: Float,
+    ) {
         Render.sprite(
             matrices,
-            x = 0,
-            y = 0,
+            x = 0f,
+            y = 0f,
             Rack.atlas,
             if (isPressed) type.pressedSprite else type.idleSprite
         )
         if (pitchControl.value == note) {
             Render.sprite(
                 matrices,
-                x = (width / 2f - type.currentPitchSprite.width / 2f).toInt(),
+                x = width / 2f - type.currentPitchSprite.width / 2f,
                 y = height - type.currentPitchSprite.height - 2,
                 Rack.atlas,
                 type.currentPitchSprite
